+++ examples/lr4_lang_ident.py
@@ -63,7 +63,13 @@ try:
     print()
 
     print("Delete specific named loaded language recogniser instance:")
-    api_response = api_instance.language_recogniser_get_details(instance_name)
+    api_response = api_instance.language_recogniser_del(instance_name)
+    print(" type(api_response)", type(api_response))
+    print(" api_response", api_response)
+    print()
+
+    print("Vaporise specific named loaded language recogniser instance:")
+    api_response = api_instance.language_recogniser_vaporise(instance_name)
     print(" type(api_response)", type(api_response))
     print(" api_response", api_response)
     print()