+++ feersum_nlu/api/person_name_entity_extractors_api.py
@@ -634,3 +634,102 @@ class PersonNameEntityExtractorsApi(object):
             _preload_content=params.get('_preload_content', True),
             _request_timeout=params.get('_request_timeout'),
             collection_formats=collection_formats)
+
+    def person_name_entity_extractor_vaporise(self, instance_name, **kwargs):  # noqa: E501
+        """Vaporise the named model.  # noqa: E501
+
+        Permanently vaporises a model even if not trashed.  # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.person_name_entity_extractor_vaporise(instance_name, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str instance_name: The name of the model instance. (required)
+        :return: PersonNameEntityExtractorInstanceDetail
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+        kwargs['_return_http_data_only'] = True
+        if kwargs.get('async'):
+            return self.person_name_entity_extractor_vaporise_with_http_info(instance_name, **kwargs)  # noqa: E501
+        else:
+            (data) = self.person_name_entity_extractor_vaporise_with_http_info(instance_name, **kwargs)  # noqa: E501
+            return data
+
+    def person_name_entity_extractor_vaporise_with_http_info(self, instance_name, **kwargs):  # noqa: E501
+        """Vaporise the named model.  # noqa: E501
+
+        Permanently vaporises a model even if not trashed.  # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.person_name_entity_extractor_vaporise_with_http_info(instance_name, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str instance_name: The name of the model instance. (required)
+        :return: PersonNameEntityExtractorInstanceDetail
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+
+        all_params = ['instance_name']  # noqa: E501
+        all_params.append('async')
+        all_params.append('_return_http_data_only')
+        all_params.append('_preload_content')
+        all_params.append('_request_timeout')
+
+        params = locals()
+        for key, val in six.iteritems(params['kwargs']):
+            if key not in all_params:
+                raise TypeError(
+                    "Got an unexpected keyword argument '%s'"
+                    " to method person_name_entity_extractor_vaporise" % key
+                )
+            params[key] = val
+        del params['kwargs']
+        # verify the required parameter 'instance_name' is set
+        if ('instance_name' not in params or
+                params['instance_name'] is None):
+            raise ValueError("Missing the required parameter `instance_name` when calling `person_name_entity_extractor_vaporise`")  # noqa: E501
+
+        collection_formats = {}
+
+        path_params = {}
+        if 'instance_name' in params:
+            path_params['instance_name'] = params['instance_name']  # noqa: E501
+
+        query_params = []
+
+        header_params = {}
+
+        form_params = []
+        local_var_files = {}
+
+        body_params = None
+        # HTTP header `Accept`
+        header_params['Accept'] = self.api_client.select_header_accept(
+            ['application/json'])  # noqa: E501
+
+        # HTTP header `Content-Type`
+        header_params['Content-Type'] = self.api_client.select_header_content_type(  # noqa: E501
+            ['application/json'])  # noqa: E501
+
+        # Authentication setting
+        auth_settings = ['APIKeyHeader', 'APIKeyHeader_old']  # noqa: E501
+
+        return self.api_client.call_api(
+            '/person_name_entity_extractors/{instance_name}/vaporise', 'POST',
+            path_params,
+            query_params,
+            header_params,
+            body=body_params,
+            post_params=form_params,
+            files=local_var_files,
+            response_type='PersonNameEntityExtractorInstanceDetail',  # noqa: E501
+            auth_settings=auth_settings,
+            async=params.get('async'),
+            _return_http_data_only=params.get('_return_http_data_only'),
+            _preload_content=params.get('_preload_content', True),
+            _request_timeout=params.get('_request_timeout'),
+            collection_formats=collection_formats)